==> Part 1.ipynb <==
import pandas as pd 
import numpy as np

data= pd.read_csv("movie_metadata.csv")

data

#import matplotlib.pyplot as plt

#data.title_year.value_counts(dropna=False).sort_index().plot(kind= "barh",figsize=(15,16))
#plt.show()

data.columns

# recommendation will be based on these features only
data = data.loc[:,['director_name','actor_1_name','actor_2_name','actor_3_name','genres','movie_title']]

data

data.isnull().sum()

data['actor_1_name'] = data['actor_1_name'].replace(np.nan, 'unknown')
data['actor_2_name'] = data['actor_2_name'].replace(np.nan, 'unknown')
data['actor_3_name'] = data['actor_3_name'].replace(np.nan, 'unknown')
data['director_name'] = data['director_name'].replace(np.nan, 'unknown')

data.isnull().sum()

data['genres'] = data['genres'].str.replace('|', ' ')

data['movie_title'] = data['movie_title'].str.lower()

# null terminating char at the end
data['movie_title'][5]

# removing the null terminating char at the end
data['movie_title'] = data['movie_title'].apply(lambda x : x[:-1])

data['movie_title'][1]

data.to_csv('data.csv',index=False)

d2= pd.read_csv("data.csv")

d2
